# logical_count: improve memory usage