--- a/test/command/suite/sharding/logical_count/condition/no_range.expected
+++ b/test/command/suite/sharding/logical_count/condition/no_range.expected
@@ -36,5 +36,27 @@ load --table Logs_20150205
 {"timestamp": "2015-02-05 13:51:00", "message": "Shutdown"}
 ]
 [[0,0.0,0.0],3]
+log_level --level dump
+[[0,0.0,0.0],true]
 logical_count Logs timestamp --filter 'message == "Shutdown"'
 [[0,0.0,0.0],2]
+#|-| [obj][open] <263>(<Logs_20150203.timestamp>):<64>(<column:fix_size>)
+#|-| [obj][open] <262>(<Logs_20150203>):<51>(<table:no_key>)
+#|-| [obj][open] <266>(<Logs_20150204.timestamp>):<64>(<column:fix_size>)
+#|-| [obj][open] <265>(<Logs_20150204>):<51>(<table:no_key>)
+#|-| [obj][open] <269>(<Logs_20150205.timestamp>):<64>(<column:fix_size>)
+#|-| [obj][open] <268>(<Logs_20150205>):<51>(<table:no_key>)
+#|-| [obj][open] <264>(<Logs_20150203.message>):<65>(<column:var_size>)
+#|-| [obj][open] <267>(<Logs_20150204.message>):<65>(<column:var_size>)
+#|-| [obj][open] <270>(<Logs_20150205.message>):<65>(<column:var_size>)
+#|-| [obj][close] <264>(<Logs_20150203.message>):<65>(<column:var_size>)
+#|-| [obj][close] <267>(<Logs_20150204.message>):<65>(<column:var_size>)
+#|-| [obj][close] <270>(<Logs_20150205.message>):<65>(<column:var_size>)
+#|-| [obj][close] <262>(<Logs_20150203>):<51>(<table:no_key>)
+#|-| [obj][close] <263>(<Logs_20150203.timestamp>):<64>(<column:fix_size>)
+#|-| [obj][close] <265>(<Logs_20150204>):<51>(<table:no_key>)
+#|-| [obj][close] <266>(<Logs_20150204.timestamp>):<64>(<column:fix_size>)
+#|-| [obj][close] <268>(<Logs_20150205>):<51>(<table:no_key>)
+#|-| [obj][close] <269>(<Logs_20150205.timestamp>):<64>(<column:fix_size>)
+log_level --level notice
+[[0,0.0,0.0],true]