--- a/test/command/suite/sharding/logical_count/condition/no_range.test
+++ b/test/command/suite/sharding/logical_count/condition/no_range.test
@@ -1,3 +1,4 @@
+#$GRN_ENABLE_REFERENCE_COUNT=yes
 #@on-error omit
 plugin_register sharding
 #@on-error default
@@ -32,4 +33,10 @@ load --table Logs_20150205
 {"timestamp": "2015-02-05 13:51:00", "message": "Shutdown"}
 ]
 
+#@add-important-log-levels dump
+#@add-ignore-log-pattern /\A\[io\]/
+log_level --level dump
 logical_count Logs timestamp --filter 'message == "Shutdown"'
+log_level --level notice
+#@remove-ignore-log-pattern /\A\[io\]/
+#@remove-important-log-levels dump