--- a/test/command/suite/sharding/logical_count/log/cover_type_all.expected
+++ b/test/command/suite/sharding/logical_count/log/cover_type_all.expected
@@ -10,7 +10,7 @@ column_create Times_20150709 timestamp_index COLUMN_INDEX Logs_20150709 timestam
 [[0,0.0,0.0],true]
 load --table Logs_20150709
 [
-{"timestamp": "2015-02-05 13:49:00"}
+{"timestamp": "2015-07-09 20:00:00"}
 ]
 [[0,0.0,0.0],1]
 log_level --level debug