--- a/test/command/suite/sharding/logical_count/log/cover_type_all.test
+++ b/test/command/suite/sharding/logical_count/log/cover_type_all.test
@@ -6,6 +6,10 @@ table_create Logs_20150709 TABLE_NO_KEY
 column_create Logs_20150709 timestamp COLUMN_SCALAR Time
 table_create Times_20150709 TABLE_PAT_KEY Time
 column_create Times_20150709 timestamp_index COLUMN_INDEX Logs_20150709 timestamp
+load --table Logs_20150709
+[
+{"timestamp": "2015-02-05 13:49:00"}
+]
 
 log_level --level debug
 #@add-important-log-levels debug