… test/command/suite/sharding/logical_count/log/select.test
… test/command/suite/sharding/logical_count/no_condition/one_physical_table.expected
… test/command/suite/sharding/logical_count/no_condition/one_physical_table.test
… test/command/suite/sharding/logical_count/no_condition/range/index/min_exclude_max_exclude.expected
… test/command/suite/sharding/logical_count/no_condition/range/index/min_exclude_max_exclude.test
… test/command/suite/sharding/logical_count/no_condition/range/index/min_include_max_include.expected
… test/command/suite/sharding/logical_count/no_condition/range/index/min_include_max_include.test
… test/command/suite/sharding/logical_count/post_filter/no_filter.expected
… test/command/suite/sharding/logical_count/post_filter/no_filter.test