Overall, I have changed `logical_count` to immediately `unref` shards or `close` temporary objects that are no longer referred like `logical_range_filter`.



TODO:

- [x] Refactor
  * Reconsider name of method, class and so on
  * Separate common method with `logical_range_filter` into a common class
    * e.g. `detect_time_classify_types`
    * If they have not match in common, I will skip this.
- [x] More test
    * Especially performance test