@HashidaTKS 2022-03-31
Performance test results.

Condition:
* 10M record per shard.
* Search 30 shards. 
* Filter with indexed columns

Sorry for my lack of details but I cannot write detailed logs or a schema and so on because it is containing private information...

**case: `GRN_ENABLE_REFERENCE_COUNT=yes`**

Execution time
|Condition|Execution time|
|--|--|
|After changed|47 sec|
|Before changed|54 sec|

Memory usage
![image](https://user-images.githubusercontent.com/15982708/161001121-29879e24-f846-4dbb-8ef3-22251ff86a68.png)

**case: `GRN_ENABLE_REFERENCE_COUNT=no`**
Execution time
|Condition|Execution time|
|--|--|
|After changed|42 sec|
|Before changed|42 sec|

Memory usage

![image](https://user-images.githubusercontent.com/15982708/161004057-179f7541-e3d6-4ffd-ab7c-9b14fb604582.png)